rename ChooseVoteOption to VotePreferenceWrapper
