--- app/components/views/ProposalDetailsPage/helpers/ChooseVoteOption/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ChooseVoteOption.jsx";
