+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/VotePreference/VotedCheckmark/VotedCheckmark.jsx
