+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/VotePreference/VotedCheckmark/VotedCheckmark.module.css
