+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/VotePreference/VotedCheckmark/index.js
