+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/VotePreferenceWrapper.module.css
@@ -1,4 +1,4 @@
-.voteChoice {
+.stakeyWrapper {
   margin-top: 15px;
   height: 64px;
 }
