+++ app/components/views/ProposalDetailsPage/helpers/VotePreferenceWrapper/index.js
@@ -1,0 +1,1 @@
+export { default } from "./VotePreferenceWrapper";
